More wind-river reviews
They really disrepected the official UCR premiere the first time in history,
and set no dead-line. Last year wth BFG out of comp. ok, but this year
with the Haneke, Lanthimos and the Sheridan. Not counting the Incovnenient Sequel.
Such a disrespect for a major festical normally only happens at the Berlinale